requirements.txt
colorama==0.4.6
langchain==0.1.12
langchain_experimental==0.0.54
langchain_openai==0.0.8
pandas==2.2.1
python-dotenv==1.0.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
gradio==4.20.0